refactor(ui): split the terminal model into runner-terminal, modularize the app shell (0046 M2)
New crates/runner-terminal — the terminal half of the Zed-style
terminal/terminal_view split: TerminalSession + TerminalBridge, palette,
fixtures + replay harness (corpus moves with the model), and a new
mappings module consolidating keystroke/paste encoding as pure
functions. No gpui dependency; the PTY stays owned by SessionManager
(the documented deviation from Zed). The deferred alacritty.rs API
isolation is noted in the 0046 spec.

runner-native's main.rs (1,405 lines) splits into chat.rs (session
lifecycle + input routing), sidebar.rs, and panes.rs, leaving a 291-line
app shell — the conflict-free lanes M3's parallel missions work in.
