+++ crates/runner-native/src/lib.rs
@@ -1,7 +1,3 @@
 pub mod bootstrap;
-pub mod fixtures;
-pub mod palette;
 pub mod pane_layout;
-pub mod replay;
-pub mod terminal;
 pub mod text_util;
