+++ crates/runner-terminal/fixtures/claude-session.ndjson
