+++ crates/runner-terminal/fixtures/claude-session.snapshot.txt
