+++ crates/runner-terminal/fixtures/width-torture.ndjson
