+++ crates/runner-terminal/src/palette.rs
